perf: job-analysis部分代码优化 #3322
1. 重构部分代码，补充部分单元测试。

+++ src/backend/job-analysis/api-job-analysis/src/main/java/com/tencent/bk/job/analysis/model/web/CommonStatisticWithRateVO.java
@@ -52,9 +52,9 @@ public class CommonStatisticWithRateVO {
     private Float momRate;
 
     @ApiModelProperty("同比趋势：1：上升，0：不变，-1：下降")
-    private Long yoyTrend;
+    private Integer yoyTrend;
 
     @ApiModelProperty("环比趋势：1：上升，0：不变，-1：下降")
-    private Long momTrend;
+    private Integer momTrend;
 
 }
